requirements.txt
# =============================================
# 📦 REQUIREMENTS.TXT - GESTOR DOCUMENTAL
# =============================================
# Actualizado: 16 de Febrero 2026
# Python: 3.8+ (Recomendado: 3.14+)
# Sistema: Windows (compatible con Linux/Mac)

# ===== CORE FRAMEWORK =====
flask==3.1.2
werkzeug==3.1.5
blinker==1.9.0
click==8.3.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3

# ===== BASE DE DATOS =====
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.11
sqlalchemy==2.0.46
greenlet==3.3.1
typing-extensions==4.15.0

# ===== SEGURIDAD =====
flask-bcrypt==1.0.1
bcrypt==5.0.0
flask-cors==6.0.2

# ===== EMAIL =====
flask-mail==0.10.0

# ===== VARIABLES DE ENTORNO =====
python-dotenv==1.2.1

# ===== HTTP REQUESTS =====
requests==2.32.3

# ===== EXCEL Y ARCHIVOS =====
openpyxl==3.1.5
et-xmlfile==2.0.0
pandas>=3.0.0
xlsxwriter==3.2.0

# ===== PROCESAMIENTO DE DATOS DE ALTO RENDIMIENTO =====
# ⭐ CRÍTICO: Para módulo DIAN vs ERP
polars==1.38.1
polars-runtime-32==1.38.1
pyarrow==23.0.1

# ===== PDF =====
fpdf2==2.8.1
reportlab==4.2.5
PyPDF2==3.0.1

# ===== SERVIDOR DE PRODUCCIÓN =====
gunicorn==23.0.0

# ===== PROGRAMACIÓN DE TAREAS =====
APScheduler==3.11.2
schedule==1.2.2
tzlocal==5.3.1
tzdata==2025.3

# ===== UTILIDADES DE FECHA Y HORA =====
python-dateutil==2.9.0.post0
pytz==2025.2

# ===== PROCESAMIENTO DE DATOS =====
numpy==2.2.3

# ===== MONITOREO Y SISTEMA =====
psutil==6.1.1

# ===== LOGGING AVANZADO =====
colorlog==6.9.0
colorama==0.4.6

# ===== VALIDACIÓN DE DATOS =====
email-validator==2.2.0

# ===== TESTING (OPCIONAL) =====
pytest==8.3.5
pytest-flask==1.3.0

# ===== FORMATOS DE SERIALIZACIÓN =====
python-multipart==0.0.20

# ===== CACHÉ (OPCIONAL) =====
flask-caching==2.3.0

# ===== RATE LIMITING (OPCIONAL - Seguridad Futura) =====
flask-limiter==3.8.0

# ===== NOTAS IMPORTANTES =====
# ⚠️  INSTALACIÓN EN WINDOWS:
#     - pyarrow y polars requieren Microsoft Visual C++ 14.0+ (Build Tools)
#     - Descargar de: https://visualstudio.microsoft.com/visual-cpp-build-tools/
#     - Instalar componente: "Herramientas de compilación de C++"
#
# 📝 COMANDOS:
#     pip install -r requirements.txt           # Instalar todo
#     pip install --upgrade -r requirements.txt # Actualizar versiones
#     pip list --outdated                       # Ver paquetes desactualizados
#
# 🔧 COMPATIBILIDAD:
#     - Python 3.8 a 3.14 compatible
#     - Versiones actualizadas a febrero 2026
#     - Todas las versiones probadas en Windows 10/11
